refactor(web)!: deprecate `Language.query` in favor of `new Query`
